Pulling lastest changes

+++ CRM/Core/BAO/OptionValue.php
@@ -240,9 +240,13 @@ class CRM_Core_BAO_OptionValue extends CRM_Core_DAO_OptionValue {
   public static function del($optionValueId) {
     $optionValue = new CRM_Core_DAO_OptionValue();
     $optionValue->id = $optionValueId;
+    if (!$optionValue->find()) {
+      return FALSE;
+    }
     if (self::updateRecords($optionValueId, CRM_Core_Action::DELETE)) {
       CRM_Core_PseudoConstant::flush();
-      return $optionValue->delete();
+      $optionValue->delete();
+      return TRUE;
     }
     return FALSE;
   }
@@ -529,7 +533,7 @@ class CRM_Core_BAO_OptionValue extends CRM_Core_DAO_OptionValue {
    */
   public static function ensureOptionValueExists($params) {
     $existingValues = civicrm_api3('OptionValue', 'get', array(
-      'option_group_name' => $params['option_group_id'],
+      'option_group_id' => $params['option_group_id'],
       'name' => $params['name'],
     ));
     if (!$existingValues['count']) {
